snapreply/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.0
pydantic-settings==2.3.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
anthropic==0.30.0
twilio==9.2.0
stripe==10.3.0
resend==2.3.0
apscheduler==3.10.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
pytz==2024.1
slowapi==0.1.9
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.126.0
